Projeto Final/requirements.txt
djangorestframework
django

